requirements.txt
aliyun-python-sdk-alidns==2.6.32
aliyun-python-sdk-core==2.13.36
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
idna==3.3
jmespath==0.10.0
pycparser==2.21
PyYAML==6.0
requests==2.28.0
urllib3==1.26.9
